refactored blog code breifly
