+++ src/main/java/com/example/springblog/models/UserRole.java
@@ -1,0 +1,17 @@
+package com.example.springblog.models;
+
+import javax.persistence.*;
+
+@Entity
+@Table(name = "user_roles")
+public class UserRole {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long id;
+
+    @Column(name = "user_id")
+    private long userId;
+
+    @Column(name = "role")
+    private String role;
+}
